Migrating to CMake build. Dir structure update and more file toucups.

+++ test/SimpleLatticeWorldTests.cpp
@@ -50,7 +50,7 @@ TEST(SimpleLatticeWorld, DefaultConstructor)
 		{(double) coords.x, (double) coords.y, (coords.z == 1) ? (double) n : coords.z - 1.0};
 
 		auto neighbors = particle->GetNeighbors();
-		ASSERT_EQ(6, neighbors.size());
+		ASSERT_EQ(6, (int)neighbors.size());
 
 		for(auto& neighbor : neighbors)
 		{
